requirements.txt
fastapi==0.12.1
uvicorn==0.6.1
sqlalchemy==1.3.2
pyjwt==1.7.1
python-dotenv==0.10.1
psycopg2==2.8
alembic==1.0.8
